requirements.txt
absl-py==1.0.0
aiodns==3.1.1
aiohttp==3.9.0
aiosignal==1.3.1
anyio==3.7.1
APScheduler==3.6.3
asn1crypto==1.5.1
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
base58==2.1.1
binance-futures-connector==4.0.0
cachetools==5.3.2
ccxt==4.1.58
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
coincurve==18.0.0
colorama==0.4.6
configparser==6.0.0
cryptography==41.0.2
cytoolz==0.12.2
dateparser==1.2.0
distlib==0.3.7
dnspython==2.4.0
docopt==0.6.2
emoji==0.4.0
eth-abi==4.1.0
eth-hash==0.5.2
eth-typing==3.4.0
eth-utils==2.2.0
exceptiongroup==1.1.2
filelock==3.13.1
flatbuffers==1.12
frozenlist==1.4.0
future==0.18.3
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h11==0.14.0
h5py==3.7.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==4.11.4
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
marshmallow==3.14.1
motor==2.5.1
multidict==6.0.4
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
parsimonious==0.9.0
pipreqs==0.4.13
platformdirs==3.11.0
pluggy==1.3.0
protobuf==3.19.4
pyaes==1.6.1
pyasn1-modules==0.2.8
pycares==4.4.0
pycparser==2.21
pycryptodome==3.18.0
pymongo==3.12.3
pyparsing==3.0.9
pyproject-api==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==20.4
pytz==2023.3.post1
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.1.0
telegram==0.0.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
TgCrypto==1.2.3
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tox==4.11.3
tronpy==0.4.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
umongo==3.0.1
urllib3==1.26.18
virtualenv==20.24.6
websocket-client==1.6.4
websockets==12.0
Werkzeug==2.1.2
wrapt==1.14.1
yarg==0.1.9
yarl==1.9.2
zipp==3.8.0
